Divided example into separate modules
